refactor: make content projection a separate service
