+++ docs/en/UI/Angular/Dom-Insertion-Service.md
@@ -1,4 +1,4 @@
-# How to Insert Scripts and Styles
+# Dom Insertion (of Scripts and Styles)
 
 You can use the `DomInsertionService` in @abp/ng.core package in order to insert scripts and styles in an easy and explicit way.
 
@@ -71,45 +71,17 @@ In the example above, `<style>body {margin: 0;}</style>` element will place at t
 
 Please refer to [ContentStrategy](./Content-Strategy.md) to see all available content strategies and how you can build your own content strategy.
 
-### How to Project Components & Templates
-
-If you pass a `ProjectionStrategy` as the first parameter of `projectContent` method, the `DomInsertionService` will resolve the projected component or template and place it at the designated target, such as containers or document body. If provided, it will also pass the component or the template a context.
-
-```js
-const componentRef = this.domInsertionService.projectContent(
-  PROJECTION_STRATEGY.AppendComponentToBody(SomeOverlayComponent)
-);
-```
-
-In the example above, `SomeOverlayComponent` component will placed at the **end** of `<body>` and a `ComponentRef` will be returned.
-
-> You should keep the returned `ComponentRef` instance, as it is a reference to the projected component and you will need that reference to destroy the projected view and the component instance.
-
-```js
-const componentRef = this.domInsertionService.projectContent(
-  PROJECTION_STRATEGY.ProjectComponentToContainer(
-    SomeOverlayComponent,
-    viewContainerRefOfTarget,
-    { someProp: "SOME_VALUE" }
-  )
-);
-```
-
-In this example, the `viewContainerRefOfTarget`, which is a `ViewContainerRef` instance, will be cleared and `SomeOverlayComponent` component will placed inside it. Moreover, the given context will be applied, so `someProp` of the component will be set to `SOME_VALUE`.
-
-Please refer to [ProjectionStrategy](./Projection-Strategy.md) to see all available projection strategies and how you can build your own projection strategy.
-
 ## API
 
 ### insertContent
 
 ```js
-injectContent(injector: Injector): ComponentRef<C> | EmbeddedViewRef<C>
+insertContent(contentStrategy: ContentStrategy): void
 ```
 
-`injector` parameter is the `Injector` instance you can pass to the projected content. It is not used in `TemplateProjectionStrategy`.
+- `contentStrategy` parameter is the primary focus here and is explained above.
 
 
 ## What's Next?
 
-- [TrackByService](./Track-By-Service.md)
+- [ContentProjectionService](./Content-Projection-Service.md)
