+++ npm/ng-packs/packages/core/src/lib/services/index.ts
@@ -1,6 +1,7 @@
 export * from './application-configuration.service';
 export * from './auth.service';
 export * from './config-state.service';
+export * from './content-projection.service';
 export * from './dom-insertion.service';
 export * from './lazy-load.service';
 export * from './localization.service';
